+++ sdk/src/types/rest.ts
@@ -1,8 +1,8 @@
+export type HttpMethod = 'GET' | 'SEARCH' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
+
 export interface RestCommand<_Input extends object, _Output extends object | unknown, _Schema extends object> {
 	(): RestRequestOptions;
 }
-
-export type HttpMethod = 'GET' | 'SEARCH' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
 
 export interface RestRequestOptions {
 	method: HttpMethod;
